newsData moved to separete file
