+++ data/newsData.ts
@@ -1,0 +1,77 @@
+
+const newsData = [
+  {
+    title: 'AI breakthrough in NLP sets new benchmarks',
+    description:
+      `A new AI model has significantly outperformed previous systems in natural language processing tasks.A new AI model has significantly outperformed previous systems in natural language processing tasks.`,
+    time: '3h ago',
+    image: require('../assets/icon.png'),
+    tag: 'AI',
+  },
+  {
+    title: 'New AI model beats humans in image recognition',
+    description:
+      'Researchers unveil an AI system that surpasses human performance in identifying images.',
+    time: '3h ago',
+    image: require('../assets/splash-icon.png'),
+    tag: 'AI',
+  },
+  {
+    title: 'Advances in neural networks drive AI technology',
+    description:
+      'Ongoing research in neural networks is pushing the boundaries of what AI can achieve.',
+    time: '3h ago',
+    tag: 'AI',
+    image: require('../assets/adaptive-icon.png'),
+  },
+  {
+    title: 'AI breakthrough in NLP sets new benchmarks',
+    description:
+      'A new AI model has significantly outperformed previous systems in natural language processing tasks.',
+    time: '3h ago',
+    image: require('../assets/icon.png'),
+    tag: 'AI',
+  },
+  {
+    title: 'New AI model beats humans in image recognition',
+    description:
+      'Researchers unveil an AI system that surpasses human performance in identifying images.',
+    time: '3h ago',
+    image: require('../assets/splash-icon.png'),
+    tag: 'AI',
+  },
+  {
+    title: 'Advances in neural networks drive AI technology',
+    description:
+      'Ongoing research in neural networks is pushing the boundaries of what AI can achieve.',
+    time: '3h ago',
+    tag: 'AI',
+    image: require('../assets/adaptive-icon.png'),
+  },
+  {
+    title: 'AI breakthrough in NLP sets new benchmarks',
+    description:
+      'A new AI model has significantly outperformed previous systems in natural language processing tasks.',
+    time: '3h ago',
+    image: require('../assets/icon.png'),
+    tag: 'AI',
+  },
+  {
+    title: 'New AI model beats humans in image recognition',
+    description:
+      'Researchers unveil an AI system that surpasses human performance in identifying images.',
+    time: '3h ago',
+    image: require('../assets/splash-icon.png'),
+    tag: 'AI',
+  },
+  {
+    title: 'Advances in neural networks drive AI technology',
+    description:
+      'Ongoing research in neural networks is pushing the boundaries of what AI can achieve.',
+    time: '3h ago',
+    tag: 'AI',
+    image: require('../assets/adaptive-icon.png'),
+  },
+];
+
+export default newsData;
